+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Whenever `minor` is bumped in the `major.minor.patch` version (e.g. `1.0.2 -> 1.
   - Dashboard now asks where to save `progress.json` and `annotations.json`.
   - Better telegraphing of span boundaries.
   - Various frontend bugfixes
+  - Allow for campaign files to contain multiple campaigns in a list
 - v1.1.4
   - Make span toolboxes stick for 0.5 second after selection
 - v1.1.3